Refactored actor and critic locations
